durations, working?

+++ dns/server/lookup_event.cpp
@@ -11,10 +11,10 @@ namespace Dometer::Dns::Server {
     LookupEvent::LookupEvent(
             const Packet& query,
             const expected<Packet, Error>& reply,
-            const std::chrono::duration<double, std::micro> duration)
+            const std::chrono::duration<float, std::micro> duration)
         :   query(query), reply(reply), duration(duration) {}
 
-    const std::chrono::duration<double, std::micro> LookupEvent::getDuration() const {
+    const std::chrono::duration<float, std::micro> LookupEvent::getDuration() const {
         return duration;
     }
